Crate accesskit_atspi_common

Modules

action 🔒

adapter 🔒

callback 🔒

context 🔒

error 🔒

events 🔒

filters 🔒

node 🔒

rect 🔒

text_attributes 🔒

util 🔒

Structs

Action

ActionHandlerWrapper

This is an implementation detail of accesskit_unix, required for robust state transitions with minimal overhead.

Adapter

AppContext

InterfaceSet

A collection type which encodes the AT-SPI interfaces an accessible object has implemented.

NodeId

PlatformNode

PlatformRoot

Rect

StateSet

The bitflag representation of all states an object may have.

WindowBounds

Enums

CoordType

The coordinate type encodes the frame of reference.

Error

Event

Granularity

Level of granularity to get text of, in relation to a cursor position.

Layer

Indicates relative stacking order of a atspi_proxies::component::ComponentProxy with respect to the onscreen visual representation of the UI.

NodeIdOrRoot

ObjectEvent

Property

RelationType

Describes a relationship between one object and another.

Role

An accessible object role.

ScrollType

Enumeration used by interface the crate::interface::Interface::Accessible to specify where an object should be placed on the screen when using scroll_to.

State

Used by various interfaces indicating every possible state of an accessibility object.

WindowEvent

Traits

ActionHandlerNoMut

This is an implementation detail of accesskit_unix, required for robust state transitions with minimal overhead.

AdapterCallback

Functions

next_adapter_id

If you use this function, you must ensure that only one adapter at a time has a given ID.

Type Aliases

Result