accesskit_atspi_common

Module adapter

Structs

Adapter

AdapterChangeHandler 🔒

Statics

NEXT_ADAPTER_ID 🔒

Functions

next_adapter_id

If you use this function, you must ensure that only one adapter at a time has a given ID.

root_window 🔒