accesskit_atspi_common::adapter

Struct Adapter

pub struct Adapter {
    id: usize,
    callback: Box<dyn AdapterCallback + Send + Sync>,
    context: Arc<Context>,
}

Fields
id: usizecallback: Box<dyn AdapterCallback + Send + Sync>context: Arc<Context>
Implementations

impl Adapter

pub fn new( app_context: &Arc<RwLock<AppContext>>, callback: impl 'static + AdapterCallback + Send + Sync, initial_state: TreeUpdate, is_window_focused: bool, root_window_bounds: WindowBounds, action_handler: impl 'static + ActionHandler + Send, ) -> Self

pub fn with_id( id: usize, app_context: &Arc<RwLock<AppContext>>, callback: impl 'static + AdapterCallback + Send + Sync, initial_state: TreeUpdate, is_window_focused: bool, root_window_bounds: WindowBounds, action_handler: impl 'static + ActionHandler + Send, ) -> Self

pub fn with_wrapped_action_handler( id: usize, app_context: &Arc<RwLock<AppContext>>, callback: impl 'static + AdapterCallback + Send + Sync, initial_state: TreeUpdate, is_window_focused: bool, root_window_bounds: WindowBounds, action_handler: Arc<dyn ActionHandlerNoMut + Send + Sync>, ) -> Self

This is an implementation detail of accesskit_unix, required for robust state transitions with minimal overhead.

fn register_tree(&self)

pub fn platform_node(&self, id: NodeId) -> PlatformNode

pub fn root_id(&self) -> NodeId

pub fn platform_root(&self) -> PlatformRoot

fn register_interfaces(&self, id: NodeId, new_interfaces: InterfaceSet)

fn unregister_interfaces(&self, id: NodeId, old_interfaces: InterfaceSet)

pub(crate) fn emit_object_event(&self, target: NodeId, event: ObjectEvent)

fn emit_root_object_event(&self, event: ObjectEvent)

pub fn set_root_window_bounds(&mut self, new_bounds: WindowBounds)

pub fn update(&mut self, update: TreeUpdate)

pub fn update_window_focus_state(&mut self, is_focused: bool)

fn window_created(&self, adapter_index: usize, window: NodeId)

fn window_activated(&self, window: &NodeWrapper<'_>)

fn window_deactivated(&self, window: &NodeWrapper<'_>)

fn window_destroyed(&self, window: NodeId)

pub fn id(&self) -> usize

pub fn is_window_focused(&self) -> bool

pub fn root_window_bounds(&self) -> WindowBounds

pub fn wrapped_action_handler( &self, ) -> Arc<dyn ActionHandlerNoMut + Send + Sync>

This is an implementation detail of accesskit_unix, required for robust state transitions with minimal overhead.