accesskit_atspi_common

Enum Role

#[repr(u32)]
pub enum Role {

    Invalid = 0,
    AcceleratorLabel = 1,
    Alert = 2,
    Animation = 3,
    Arrow = 4,
    Calendar = 5,
    Canvas = 6,
    CheckBox = 7,
    CheckMenuItem = 8,
    ColorChooser = 9,
    ColumnHeader = 10,
    ComboBox = 11,
    DateEditor = 12,
    DesktopIcon = 13,
    DesktopFrame = 14,
    Dial = 15,
    Dialog = 16,
    DirectoryPane = 17,
    DrawingArea = 18,
    FileChooser = 19,
    Filler = 20,
    FocusTraversable = 21,
    FontChooser = 22,
    Frame = 23,
    GlassPane = 24,
    HTMLContainer = 25,
    Icon = 26,
    Image = 27,
    InternalFrame = 28,
    Label = 29,
    LayeredPane = 30,
    List = 31,
    ListItem = 32,
    Menu = 33,
    MenuBar = 34,
    MenuItem = 35,
    OptionPane = 36,
    PageTab = 37,
    PageTabList = 38,
    Panel = 39,
    PasswordText = 40,
    PopupMenu = 41,
    ProgressBar = 42,
    PushButton = 43,
    RadioButton = 44,
    RadioMenuItem = 45,
    RootPane = 46,
    RowHeader = 47,
    ScrollBar = 48,
    ScrollPane = 49,
    Separator = 50,
    Slider = 51,
    SpinButton = 52,
    SplitPane = 53,
    StatusBar = 54,
    Table = 55,
    TableCell = 56,
    TableColumnHeader = 57,
    TableRowHeader = 58,
    TearoffMenuItem = 59,
    Terminal = 60,
    Text = 61,
    ToggleButton = 62,
    ToolBar = 63,
    ToolTip = 64,
    Tree = 65,
    TreeTable = 66,
    Unknown = 67,
    Viewport = 68,
    Window = 69,
    Extended = 70,
    Header = 71,
    Footer = 72,
    Paragraph = 73,
    Ruler = 74,
    Application = 75,
    Autocomplete = 76,
    Editbar = 77,
    Embedded = 78,
    Entry = 79,
    CHART = 80,
    Caption = 81,
    DocumentFrame = 82,
    Heading = 83,
    Page = 84,
    Section = 85,
    RedundantObject = 86,
    Form = 87,
    Link = 88,
    InputMethodWindow = 89,
    TableRow = 90,
    TreeItem = 91,
    DocumentSpreadsheet = 92,
    DocumentPresentation = 93,
    DocumentText = 94,
    DocumentWeb = 95,
    DocumentEmail = 96,
    Comment = 97,
    ListBox = 98,
    Grouping = 99,
    ImageMap = 100,
    Notification = 101,
    InfoBar = 102,
    LevelBar = 103,
    TitleBar = 104,
    BlockQuote = 105,
    Audio = 106,
    Video = 107,
    Definition = 108,
    Article = 109,
    Landmark = 110,
    Log = 111,
    Marquee = 112,
    Math = 113,
    Rating = 114,
    Timer = 115,
    Static = 116,
    MathFraction = 117,
    MathRoot = 118,
    Subscript = 119,
    Superscript = 120,
    DescriptionList = 121,
    DescriptionTerm = 122,
    DescriptionValue = 123,
    Footnote = 124,
    ContentDeletion = 125,
    ContentInsertion = 126,
    Mark = 127,
    Suggestion = 128,
    PushButtonMenu = 129,

}

An accessible object role. To think of it in terms of HTML, any semantic element likely has a corollary in this enum. For example: <button>, <input>, <form> or <h4>. Non-semantic elements like <span>, <div> and <b> will not be represented here, and this information is not passed through via the atspi library. TODO: add examples for GTK/Qt libraries in addition to HTML examples.

Variants

Invalid = 0

A role indicating an error condition, such as uninitialized Role data, or an error deserializing.

AcceleratorLabel = 1

Object is a label indicating the keyboard accelerators for the parent.

Alert = 2

Object is used to alert the user about something.

Animation = 3

Object contains a dynamic or moving image of some kind.

Arrow = 4

Object is a 2d directional indicator.

Calendar = 5

Object contains one or more dates, usually arranged into a 2d list.

Canvas = 6

Object that can be drawn into and is used to trap events.

CheckBox = 7

A choice that can be checked or unchecked and provides a separate indicator for the current state.

CheckMenuItem = 8

A menu item that behaves like a check box. See Self::CheckBox.

ColorChooser = 9

A specialized dialog that lets the user choose a color.

ColumnHeader = 10

The header for a column of data.

ComboBox = 11

A list of choices the user can select from.

DateEditor = 12

An object which allows entry of a date.

DesktopIcon = 13

An inconifed internal frame within a Role::DesktopFrame.

DesktopFrame = 14

A pane that supports internal frames and iconified versions of those internal frames.

Dial = 15

An object that allows a value to be changed via rotating a visual element, or which displays a value via such a rotating element.

Dialog = 16

A top level window with title bar and a border.

DirectoryPane = 17

A pane that allows the user to navigate through and select the contents of a directory.

DrawingArea = 18

An object used for drawing custom user interface elements.

FileChooser = 19

A specialized dialog that displays the files in the directory and lets the user select a file, browse a different directory, or specify a filename.

Filler = 20

A object that fills up space in a user interface.

FocusTraversable = 21

Don’t use, reserved for future use.

FontChooser = 22

Allows selection of a display font.

Frame = 23

A top level window with a title bar, border, menubar, etc.

GlassPane = 24

A pane that is guaranteed to be painted on top of all panes beneath it.

HTMLContainer = 25

A document container for HTML, whose children represent the document content.

Icon = 26

A small fixed size picture, typically used to decorate components.

Image = 27

An image, typically static.

InternalFrame = 28

A frame-like object that is clipped by a desktop pane.

Label = 29

An object used to present an icon or short string in an interface.

LayeredPane = 30

A specialized pane that allows its children to be drawn in layers, providing a form of stacking order.

List = 31

An object that presents a list of objects to the user and * allows the user to select one or more of them.

ListItem = 32

An object that represents an element of a list.

Menu = 33

An object usually found inside a menu bar that contains a list of actions the user can choose from.

MenuBar = 34

An object usually drawn at the top of the primary dialog box of an application that contains a list of menus the user can choose from.

MenuItem = 35

An object usually contained in a menu that presents an action the user can choose.

OptionPane = 36

A specialized pane whose primary use is inside a dialog.

PageTab = 37

An object that is a child of a page tab list.

PageTabList = 38

An object that presents a series of panels (or page tabs), one at a time,through some mechanism provided by the object.

Panel = 39

A generic container that is often used to group objects.

PasswordText = 40

A text object uses for passwords, or other places where the text content is not shown visibly to the user.

PopupMenu = 41

A temporary window that is usually used to offer the user a list of choices, and then hides when the user selects one of those choices.

ProgressBar = 42

An object used to indicate how much of a task has been completed.

PushButton = 43

An object the user can manipulate to tell the application to do something.

RadioButton = 44

A specialized check box that will cause other radio buttons in the same group to become unchecked when this one is checked.

RadioMenuItem = 45

Object is both a menu item and a “radio button”. See Self::RadioButton.

RootPane = 46

A specialized pane that has a glass pane and a layered pane as its children.

RowHeader = 47

The header for a row of data.

ScrollBar = 48

An object usually used to allow a user to incrementally view a large amount of data by moving the bounds of a viewport along a one-dimensional axis.

ScrollPane = 49

A scroll pane: the pane in which the scrollable content is contained within. An object that allows a user to incrementally view a large amount of information. Self::ScrollPane objects are usually accompanied by Self::ScrollBar controllers, on which the crate::RelationType::ControllerFor and crate::RelationType::ControlledBy reciprocal relations are set.

Separator = 50

An object usually contained in a menu to provide a visible and logical separation of the contents in a menu.

Slider = 51

An object that allows the user to select from a bounded range. Unlike Self::ScrollBar, Self::Slider objects need not control ‘viewport’-like objects.

SpinButton = 52

An object which allows one of a set of choices to be selected, and which displays the current choice.

SplitPane = 53

A specialized panel that presents two other panels at the same time.

StatusBar = 54

Object displays non-quantitative status information (c.f. Self::ProgressBar)

Table = 55

An object used to represent information in terms of rows and columns.

TableCell = 56

A ‘cell’ or discrete child within a Table. Note: Table cells need not have Self::TableCell, other crate::Role values are valid as well.

TableColumnHeader = 57

An object which labels a particular column in an Self::Table.

TableRowHeader = 58

An object which labels a particular row in a Self::Table. TableProxy rows and columns may also be labelled via the crate::RelationType::LabelFor/crate::RelationType::LabelledBy relationships. See: AccessibleProxy::get_relation_type.

TearoffMenuItem = 59

Object allows menu to be removed from menubar and shown in its own window.

Terminal = 60

An object that emulates a terminal.

Text = 61

An interactive widget that supports multiple lines of text and optionally accepts user input, but whose purpose is not to solicit user input. Thus Self::Text is appropriate for the text view in a plain text editor but inappropriate for an input field in a dialog box or web form. For widgets whose purpose is to solicit input from the user, see Self::Entry and Self::PasswordText. For generic objects which display a brief amount of textual information, see Self::Static.

ToggleButton = 62

A specialized push button that can be checked or unchecked, but does not provide a separate indicator for the current state.

ToolBar = 63

A bar or palette usually composed of push buttons or toggle buttons.

ToolTip = 64

An object that provides information about another object.

Tree = 65

An object used to repsent hierarchical information to the user.

TreeTable = 66

An object that presents both tabular and hierarchical info to the user.

Unknown = 67

When the role cannot be accurately reported, this role will be set.

Viewport = 68

An object usually used in a scroll pane, or to otherwise clip a larger object or content renderer to a specific onscreen viewport.

Window = 69

A top level window with no title or border.

Extended = 70

means that the role for this item is known, but not included in the core enumeration.

Header = 71

An object that serves as a document header.

Footer = 72

An object that serves as a document footer.

Paragraph = 73

An object which is contains a single paragraph of text content. See also Self::Text.

Ruler = 74

An object which describes margins and tab stops, etc. for text objects which it controls (should have crate::RelationType::ControllerFor relation to such).

Application = 75

An object corresponding to the toplevel accessible of an application, which may contain Self::Frame objects or other accessible objects. Children of objects with the Self::DesktopFrame role are generally Self::Application objects.

Autocomplete = 76

The object is a dialog or list containing items for insertion into an entry widget, for instance a list of words for completion of a text entry.

Editbar = 77

The object is an editable text object in a toolbar.

Embedded = 78

The object is an embedded component container. This role is a “grouping” hint that the contained objects share a context which is different from the container in which this accessible is embedded. In particular, it is used for some kinds of document embedding, and for embedding of out-of-process component, “panel applets”, etc.

Entry = 79

The object is a component whose textual content may be entered or modified by the user, provided crate::State::Editable is present. A readonly Self::Entry object (i.e. where crate::State::Editable is not present) implies a read-only ‘text field’ in a form, as opposed to a title, label, or caption.

CHART = 80

The object is a graphical depiction of quantitative data. It may contain multiple subelements whose attributes and/or description may be queried to obtain both the quantitative data and information about how the data is being presented. The crate::RelationType::LabelledBy relation is particularly important in interpreting objects of this type, as is the accessible description property. See Self::Caption.

Caption = 81

The object contains descriptive information, usually textual, about another user interface element such as a table, chart, or image.

DocumentFrame = 82

The object is a visual frame or container which contains a view of document content. Self::DocumentFrames may occur within another DocumentProxy instance, in which case the second document may be said to be embedded in the containing instance. HTML frames are often Self::DocumentFrame: Either this object, or a singleton descendant, should implement the crate::Interface::Document interface.

Heading = 83

Heading: this is a heading with a level (usually 1-6). This is represented by <h1> through <h6> in HTML. The object serves as a heading for content which follows it in a document. The ‘heading level’ of the heading, if available, may be obtained by querying the object’s attributes.

Page = 84

The object is a containing instance which encapsulates a page of information. Self::Page is used in documents and content which support a paginated navigation model.

Section = 85

The object is a containing instance of document content which constitutes a particular ‘logical’ section of the document. The type of content within a section, and the nature of the section division itself, may be obtained by querying the object’s attributes. Sections may be nested.

RedundantObject = 86

The object is redundant with another object in the hierarchy, and is exposed for purely technical reasons. Objects of this role should be ignored by clients, if they are encountered at all.

Form = 87

The object is a containing instance of document content which has within it components with which the user can interact in order to input information; i.e. the object is a container for pushbuttons, comboboxes, text input fields, and other ‘GUI’ components. Self::Form should not, in general, be used for toplevel GUI containers or dialogs, but should be reserved for ‘GUI’ containers which occur within document content, for instance within Web documents, presentations, or text documents. Unlike other GUI containers and dialogs which occur inside application instances, Self::Form containers’ components are associated with the current document, rather than the current foreground application or viewer instance.

Link = 88

The object is a hypertext anchor, i.e. a “link” in a hypertext document. Such objects are distinct from ‘inline’ content which may also use the crate::Interface::Hypertext/crate::Interface::Hyperlink interfaces to indicate the range/location within a text object where an inline or embedded object lies.

InputMethodWindow = 89

The object is a window or similar viewport which is used to allow composition or input of a ‘complex character’, in other words it is an “input method window”.

TableRow = 90

A row in a table.

TreeItem = 91

An object that represents an element of a tree.

DocumentSpreadsheet = 92

A document frame which contains a spreadsheet.

DocumentPresentation = 93

A document frame which contains a presentation or slide content.

DocumentText = 94

A document frame which contains textual content, such as found in a word processing application.

DocumentWeb = 95

A document frame which contains HTML or other markup suitable for display in a web browser.

DocumentEmail = 96

A document frame which contains email content to be displayed or composed either in plain text or HTML.

Comment = 97

An object found within a document and designed to present a comment, note, or other annotation. In some cases, this object might not be visible until activated.

ListBox = 98

A non-collapsible list of choices the user can select from.

Grouping = 99

A group of related widgets. This group typically has a label.

ImageMap = 100

An image map object. Usually a graphic with multiple hotspots, where each hotspot can be activated resulting in the loading of another document or section of a document.

Notification = 101

A transitory object designed to present a message to the user, typically at the desktop level rather than inside a particular application.

InfoBar = 102

An object designed to present a message to the user within an existing window.

LevelBar = 103

A bar that serves as a level indicator to, for instance, show the strength of a password or the state of a battery.

TitleBar = 104

A bar that serves as the title of a window or a dialog.

BlockQuote = 105

An object which contains a text section that is quoted from another source.

Audio = 106

An object which represents an audio element.

Video = 107

An object which represents a video element.

Definition = 108

A definition of a term or concept.

Article = 109

A section of a page that consists of a composition that forms an independent part of a document, page, or site. Examples: A blog entry, a news story, a forum post.

Landmark = 110

A region of a web page intended as a navigational landmark. This is designed to allow Assistive Technologies to provide quick navigation among key regions within a document.

Log = 111

A text widget or container holding log content, such as chat history and error logs. In this role there is a relationship between the arrival of new items in the log and the reading order. The log contains a meaningful sequence and new information is added only to the end of the log, not at arbitrary points.

Marquee = 112

A container where non-essential information changes frequently. Common usages of marquee include stock tickers and ad banners. The primary difference between a marquee and a log is that logs usually have a meaningful order or sequence of important content changes.

Math = 113

A text widget or container that holds a mathematical expression.

Rating = 114

A rating system, generally out of five stars, but it does not need to be that way. There is no tag nor role for this in HTML, however. A widget whose purpose is to display a rating, such as the number of stars associated with a song in a media player. Objects of this role should also implement crate::Interface::Value.

Timer = 115

An object containing a numerical counter which indicates an amount of elapsed time from a start point, or the time remaining until an end point.

Static = 116

A generic non-container object whose purpose is to display a brief amount of information to the user and whose role is known by the implementor but lacks semantic value for the user. Examples in which Self::Static is appropriate include the message displayed in a message box and an image used as an alternative means to display text. Self::Static should not be applied to widgets which are traditionally interactive, objects which display a significant amount of content, or any object which has an accessible relation pointing to another object. The displayed information, as a general rule, should be exposed through the accessible name of the object. For labels which describe another widget, see Self::Label. For text views, see Self::Text. For generic containers, see Self::Panel. For objects whose role is not known by the implementor, see Self::Unknown.

MathFraction = 117

An object that represents a mathematical fraction.

MathRoot = 118

An object that represents a mathematical expression displayed with a radical.

Subscript = 119

An object that contains text that is displayed as a subscript.

Superscript = 120

An object that contains text that is displayed as a superscript.

DescriptionList = 121

An object that represents a list of term-value groups. A term-value group represents an individual description and consist of one or more names (Self::DescriptionTerm) followed by one or more values (Self::DescriptionValue). For each list, there should not be more than one group with the same term name.

DescriptionTerm = 122

An object that represents a term or phrase with a corresponding definition.

DescriptionValue = 123

An object that represents the description, definition, or value of a term.

Footnote = 124

An object that contains the text of a footnote.

ContentDeletion = 125

Content previously deleted or proposed to be deleted, e.g. in revision history or a content view providing suggestions from reviewers.

ContentInsertion = 126

Content previously inserted or proposed to be inserted, e.g. in revision history or a content view providing suggestions from reviewers.

Mark = 127

A run of content that is marked or highlighted, such as for reference purposes, or to call it out as having a special purpose. If the marked content has an associated section in the document elaborating on the reason for the mark, then crate::RelationType::Details should be used on the mark to point to that associated section. In addition, the reciprocal relation crate::RelationType::DetailsFor should be used on the associated content section to point back to the mark.

Suggestion = 128

A container for content that is called out as a proposed change from the current version of the document, such as by a reviewer of the content. An object with this role should include children with Self::ContentDeletion and/or Self::ContentInsertion, in any order, to indicate what the actual change is.

PushButtonMenu = 129

A specialized push button to open a menu.

Implementations

impl Role

pub fn name(&self) -> &'static str

Get a readable, English name from the role.